A new version for pcf_flags.
Replace #2638
